requirements.txt
aiodns==2.0.0
aiohttp==3.7.4.post0
alabaster==0.7.12
asgiref==3.3.1
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.0
beautifulsoup4==4.9.3
bs4==0.0.1
ccxt==1.47.60
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
config==0.5.0.post0
cryptography==3.4.7
dataframe==0.2.1.3
Django==3.0.5
django-cors-headers==3.7.0
django-request==1.5.6
django-rest-framework==0.1.0
djangorestframework==3.12.3
djongo==1.3.4
docutils==0.17
idna==2.10
imagesize==1.2.0
importlib-metadata==3.10.0
iniconfig==1.1.1
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
nose==1.3.7
numpy==1.20.2
packaging==20.9
pandas==1.2.3
pluggy==0.13.1
py==1.10.0
pycares==3.1.1
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pyupbit==0.2.14
requests==2.25.1
scipy==1.6.2
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==3.5.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.2.4
tabulate==0.8.9
toml==0.10.2
typing-extensions==3.7.4.3
upbitpy==1.0.1
urllib3==1.26.4
websockets==8.1
yarl==1.1.0
zipp==3.4.1
